Add trialize_data.m as part of new branch to re-base analysis code
Want to re-base analysis code because the first step, i.e. trialize_neurons.m currently results in a lot of redundancy, which makes later analyses more difficult; I'm finding that some later analyses involve un-doing what was done in the first step. I think it would be better to have a larger number of more atomic steps that result in less redundancy and that will make analysis  more flexible.

trialize_data.m yields a neuron-major array of structs, each of which contains information about the parameters of every trial the neuron was observed in. Because all of the neurons in a given session were observed in all of the same trials, however, this trial parameter information is really redundant.